requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.9.1
    # via
    #   django
    #   django-cors-headers
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
django==5.2.5
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-neomodel
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
django-cors-headers==4.7.0
    # via -r requirements.in
django-neomodel==0.2.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.28.0
    # via -r requirements.in
drf-spectacular-sidecar==2025.9.1
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.25.1
    # via drf-spectacular
jsonschema-specifications==2025.4.1
    # via jsonschema
neo4j==5.19.0
    # via neomodel
neomodel==5.3.3
    # via
    #   -r requirements.in
    #   django-neomodel
pyjwt==2.10.1
    # via -r requirements.in
python-decouple==3.8
    # via -r requirements.in
pytz==2025.2
    # via neo4j
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.3
    # via django
uritemplate==4.2.0
    # via drf-spectacular
